move vmware_fusion setup hook later
i think when it was originally implemented, the vmware-fusion plugin did
not actually use the WaitForCommunicator builtin so we had to hook on
something else in the vicinity.

but depending on the box, trying to communicate via ssh right after
`Boot` does not always work because ssh is not ready yet.

fixes #79
closes #86

big thanks to @kalefranz and @Xylakant for their help in getting this
figured out!