ISSUE #249 Fixing inherited file handles
On Mac and Linux the DNS server spawn code automatically inherited
filehandles for STDIN, STDOUT, and STDERR.  This fix closes those
filehandles and sets all other filehandles to not explicitly redirected
to uninherited.

This is already the default for Windows.

This fix is thanks to @dustymabe and his research on a similiar issue
for vagrant-sshfs
https://github.com/dustymabe/vagrant-sshfs/issues/41