Improve concurrency behaviour for multi-machine envs
Use class variables to support configuration of the DNS server once per
run for multi-machine environments.

Process level locking is done using a class variable containing a Mutex
and a simple synchronization block. Subsequent threads running the
configuration will skip for items already defined/started and so should
be no-ops.

This allows a single attempt to bring up the DNS server, and ensures
that for parallel bring up with multi-machine environments that one a
single attempt is made to start the server, and therefore the correct
pid will be available during destroy to cleanup.

Fixes #304